requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.15.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
imbalanced-learn>=0.11.0
